api/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
certifi==2025.1.31
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
constantly==23.10.4
cryptography==44.0.3
daphne==4.1.2
Django==5.2
djangorestframework==3.16.0
duckdb==1.2.1
fastapi==0.115.11
filelock==3.18.0
fsspec==2025.3.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
Jinja2==3.1.6
Markdown==3.7
MarkupSafe==3.0.2
mpmath==1.3.0
msgpack==1.1.0
networkx==3.4.2
numpy==2.2.3
packaging==24.2
pandas==2.2.3
plotly==5.24.1
preswald==0.1.42
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.1
redis==6.0.0
requests==2.32.3
scipy==1.15.2
service-identity==24.2.0
setuptools==76.0.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.46.1
sympy==1.14.0
tenacity==9.0.0
toml==0.10.2
tomli==2.2.1
torch==2.7.0
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
websockets==10.4
zope.interface==7.2
